+++ js/class/Resources.js
