+++ js/util/detect.js
